backend/requirements.txt
alembic==1.3.2
autopep8==1.4.4
certifi==2019.11.28
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
